Replaces base Exception usage with custom exceptions - Closes #10; Adds basic/required validation of EntityExtensions - Closes #9; Adds XML docs to each public method - Closes #11; Updates nuspec to point to GitHub releases - Closes #5

+++ src/CloudAwesome.Xrm.Core/CloudAwesome.Xrm.Core/EntityExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-
+using CloudAwesome.Xrm.Core.Exceptions;
 using Microsoft.Crm.Sdk.Messages;
 using Microsoft.Xrm.Sdk;
 using Microsoft.Xrm.Sdk.Messages;
@@ -9,29 +9,64 @@ using Microsoft.Xrm.Sdk.Query;
 
 namespace CloudAwesome.Xrm.Core
 {
-    // TODO - * Enum.TryParse wrapper?
-
     public static class EntityExtensions
     {
+        /// <summary>
+        /// Create new record in CDS
+        /// </summary>
+        /// <param name="entity">Record to create. Can be a base entity or any early bound entity inheriting from Entity</param>
+        /// <param name="organizationService">IOrganization reference</param>
+        /// <exception cref="OperationPreventedException">Throws when Logical Name of the record is null or empty</exception>
+        /// <returns>EntityReference of the created record</returns>
         public static EntityReference Create(this Entity entity, IOrganizationService organizationService)
         {
-            // TODO - tracing (and validation?)
+            if (string.IsNullOrEmpty(entity.LogicalName))
+            {
+                throw new OperationPreventedException("Cannot create a new record if the entity's logical name is null or empty. IT is recommended to use early bound entity types if possible");
+            }
             entity.Id = organizationService.Create(entity);
             return new EntityReference(entity.LogicalName, entity.Id);
         }
 
+        /// <summary>
+        /// Delete the referenced record in CRM
+        /// </summary>
+        /// <param name="entity">Record to delete. Can be a base entity or any early bound entity inheriting from Entity</param>
+        /// <param name="organizationService">IOrganization reference</param>
+        /// <exception cref="OperationPreventedException">Throws when primary GUID of the record is null</exception>
         public static void Delete(this Entity entity, IOrganizationService organizationService)
         {
-            // TODO - tracing
+            if (entity.Id == null)
+            {
+                throw new OperationPreventedException("Cannot delete a record if GUID is null");
+            }
             organizationService.Delete(entity.LogicalName, entity.Id);
         }
 
-        public static void Update(this Entity entity, IOrganizationService organizationService)
+        /// <summary>
+        /// Update the reference record in CRM
+        /// </summary>
+        /// <param name="entity">Record to update. Can be a base entity or any early bound entity inheriting from Entity</param>
+        /// <param name="organizationService">IOrganization reference</param>
+        /// <exception cref="OperationPreventedException">Throws when primary GUID of the record is null</exception>
+        /// <returns>EntityReference of the updated record</returns>
+        public static EntityReference Update(this Entity entity, IOrganizationService organizationService)
         {
-            // TODO - tracing and validation
+            if (entity.Id == null)
+            {
+                throw new OperationPreventedException("Cannot update a record if GUID is null");
+            }
             organizationService.Update(entity);
+            return entity.ToEntityReference();
         }
 
+        /// <summary>
+        /// Update existing record or create new if doesn't already exist, based on QueryBase given
+        /// </summary>
+        /// <param name="entity">Record to create or update</param>
+        /// <param name="organizationService">IOrganization reference</param>
+        /// <param name="query">QueryBase implementation to determine if the record already exists. To update, method expects a single result to be returned</param>
+        /// <returns>EntityReference of updated or created record</returns>
         public static EntityReference CreateOrUpdate(this Entity entity, IOrganizationService organizationService,
             QueryBase query)
         {
@@ -61,17 +96,35 @@ namespace CloudAwesome.Xrm.Core
             //})).Id;
         }
 
+        /// <summary>
+        /// Retrieves an entity based on the current entity's primary ID.
+        /// If other query criteria are required used one of the QueryExtensions methods
+        /// </summary>
+        /// <param name="entity">Current record to retrieve, must have primary ID populated</param>
+        /// <param name="organizationService">IOrganization reference</param>
+        /// <param name="columnSet">ColumnSet to retrieve. Pass null to retrieve all columns (not recommended in most cases)</param>
+        /// <exception cref="OperationPreventedException">Throws when primary GUID of the record is null</exception>
+        /// <returns>Entity retrieved. Use entity.ToEntity&lt;&gt;() to parse into an early bound entity type</returns>
         public static Entity Retrieve(this Entity entity,
             IOrganizationService organizationService, ColumnSet columnSet = null)
         {
             if (Guid.Empty == entity.Id)
             {
-                throw new Exception("Cannot retrieve Entity if ID is null or empty. Try a query instead.");
+                throw new OperationPreventedException("Cannot retrieve Entity if ID is null or empty. Try a query instead.");
             }
             return organizationService.Retrieve(entity.LogicalName, entity.Id, columnSet?? new ColumnSet(true));
         }
 
-        public static void CloneFrom(this Entity targetEntity, Entity sourceEntity, List<string> excludeAttributesList = null)
+
+        /// <summary>
+        /// Creates a copy of the source entity into the target entity. Always excludes the record's GUID and can exclude other fields listed
+        /// </summary>
+        /// <param name="targetEntity">New entity receive cloned attribute values</param>
+        /// <param name="sourceEntity">Existing record to source attribute values</param>
+        /// <param name="excludeAttributesList">Optional list of attribute schema names to ignore from clone</param>
+        /// <exception cref="Exception">Throws a generic exception if the source and target entities are not of the same type (based on logical name)</exception>
+        /// <returns>Cloned Entity. Use entity.ToEntity&lt;&gt;() to parse into an early bound entity type as required</returns>
+        public static Entity CloneFrom(this Entity targetEntity, Entity sourceEntity, List<string> excludeAttributesList = null)
         {
             if (targetEntity.LogicalName != sourceEntity.LogicalName)
             {
@@ -88,6 +141,8 @@ namespace CloudAwesome.Xrm.Core
                 
                 targetEntity[attr.Key] = attr.Value;
             }
+
+            return targetEntity;
         }
 
         public static void Associate(this Entity entity, IOrganizationService organizationService,
@@ -122,11 +177,5 @@ namespace CloudAwesome.Xrm.Core
             //    new Relationship(relationshipName), relatedEntities);
         }
         
-        public static void SetState(this Entity entity, 
-            IOrganizationService organization, int stateCode, int statusCode)
-        {
-            // Has SetState now been deprecated?
-            throw FeatureRequestException.NotImplementedFeatureException(typeof(SetStateRequest).ToString());
-        }
     }
 }
